requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sentence-transformers==2.4.0
transformers>=4.41.0
torch>=2.0
accelerate==0.27.2
faiss-cpu
numpy>=1.24
pillow
python-multipart
langchain==0.1.13
pypdf
num2words>=0.5 